+++ RPiAPMCopter/receiver.h
@@ -15,14 +15,14 @@ class Device;
 class Receiver {
 private:
   char    m_cBuffer[512];
-  
+
   Device  *m_pHalBoard;
   uint32_t m_iSerialTimer;
 
   uint8_t calc_chksum     (char *);
   bool    verf_chksum     (char *str, char *chk);
   float*  parse_pid_substr(char*);
-  
+
 protected:
   bool    parse_ctrl_com  (char*);
   bool    parse_gyr_cor   (char*);
@@ -31,14 +31,14 @@ protected:
   bool    parse_pid_conf  (char*);
 
   bool    parse           (char*);
-  
+
 public:
   Receiver(Device *pHalBoard);
 
   int16_t m_pChannelsRC[APM_IOCHANNEL_COUNT];
-  // Read from serial bus 
-  bool read_uartA(uint32_t bytesAvail); // console in APM 2
-  bool read_uartC(uint32_t bytesAvail); // radio in APM 2
+  // Read from serial bus
+  bool read_uartA(uint16_t bytesAvail); // console in APM 2
+  bool read_uartC(uint16_t bytesAvail); // radio in APM 2
   // time since last command string was parsed successfully
   uint32_t  time_elapsed();
 };
